refactor: split oversized files, extract state module, harden Makefile
- Split ai_persistence.go (781→659 LOC): extract MCP handler to ai_persistence_handler.go
- Split security_diff.go: extract diff helpers to security_diff_helpers.go
- Split bridge.go: extract forwarding/error/restart to bridge_forward.go
- Extract index.ts god module: mutable state → state.ts (569→468 LOC)
- Add check-file-length as prerequisite of `make test`
- Add compile-ts as prerequisite of `make test-js`
- Update contract tests and lint-hardening for new file locations
