+++ scripts/lint-hardening.sh
@@ -348,7 +348,7 @@ fi
 # ─────────────────────────────────────────────
 bold "12. Checking bridge dispatch signal() coverage..."
 
-BRIDGE_FILE="cmd/dev-console/bridge.go"
+BRIDGE_FILE="cmd/dev-console/bridge_forward.go"
 if [ -f "$BRIDGE_FILE" ]; then
   BRIDGE_SIGNAL_ISSUES=""
   FUNC_START=$(grep -n 'func bridgeForwardRequest' "$BRIDGE_FILE" | head -1 | cut -d: -f1)
