+++ src/background/init.ts
@@ -88,8 +88,9 @@ async function initializeExtensionAsync(): Promise<void> {
     console.log('[Gasoline] Storage value:', aiPilotEnabled, '| Cache value:', index.__aiWebPilotEnabledCache)
 
     // Execute any pending pilot init callback
-    if (index.__pilotInitCallback) {
-      index.__pilotInitCallback()
+    const pilotCb = index.__pilotInitCallback
+    if (pilotCb) {
+      pilotCb()
       setPilotInitCallback(null)
     }
